+++ Changelog.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+ - Field match access
+ - Feature Email Verification
+ 
+### Changed
+ - FJS-1078: deleted repeated word in access tab description
+ - FOR-2719: Made actions disabled to uncheck in the export template
+ - FJS-1069: Data Submission table does not show the amount of submissions
+ - Remove kickbox settings
+ - Work on stages: Enable to rename Live stage, Added 2 default stages
+
+### Fixed
+ - FJS-1071: Set initial access
+
+## 7.1.15-rc.1
+### Changed
+ - Adding Group permission levels.
+ - Upgrade formiojs@4.11.2-rc.5
+
 ## 7.1.14
 ### Fixed 
  - Couple of tweaks to fix on premise remote stages.
@@ -67,6 +87,51 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 7.1.0-7.1.4
 ### Changed
  - Licensing UI
+
+[Unreleased]
+### Changed
+ - Add more license management to current production.
+
+## 7.0.39
+### Changed
+ - Upgrade "@formio/premium": "github:formio/premium#v1.14.2",
+ - Upgrade "formiojs": "^4.11.2"
+ - Upgrade "formmanager": "github:formio/formmanager#v1.94.8"
+ - Upgrade "ng-formio": "^4.15.2"
+ - Upgrade "formio-tenant": "^1.13.17"
+
+## 7.0.39-rc.5
+### Changed
+ - Reverted group permissions since that is in the 7.1.x build of portal.
+
+## 7.0.39-rc.4
+### Changed
+ - Adding group permission levels.
+
+## 7.0.39-rc.3
+### Changed
+ - Upgrade "@formio/premium": "github:formio/premium#v1.14.2-rc.7",
+ - Upgrade "formiojs": "^4.11.2-rc.4"
+ - Upgrade "formmanager": "github:formio/formmanager#v1.94.8-rc.3"
+ - Upgrade "ng-formio": "^4.15.2-rc.3"
+ - Upgrade "formio-tenant": "^1.13.17-rc.3"
+
+## 7.0.39-rc.2
+### Changed
+ - Upgrade "@formio/premium": "github:formio/premium#v1.14.2-rc.5",
+ - Upgrade "formiojs": "^4.11.2-rc.2"
+ - Upgrade "formmanager": "github:formio/formmanager#v1.94.8-rc.2"
+ - Upgrade "ng-formio": "^4.15.2-rc.2"
+ - Upgrade "formio-tenant": "^1.13.17-rc.2"
+
+## 7.0.39-rc.1
+### Changed
+ - Upgrade "@formio/premium": "github:formio/premium#v1.14.2-rc.4",
+ - Upgrade "formiojs": "^4.11.2-rc.1"
+ - Upgrade "formmanager": "github:formio/formmanager#v1.94.8-rc.1"
+ - Upgrade "ng-formio": "^4.15.2-rc.1"
+ - Upgrade "formio-tenant": "^1.13.17-rc.1"
+>>>>>>> 28f3e8e3a894493496b33971a9d60827ac4b9544
 
 ## 7.0.38
 ### Changed